vector24/requirements.txt
mss>=9.0.0
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
pyinstaller>=6.0.0
pypresence>=4.3.0
